requirements.txt
asgiref==3.7.2
astroid==3.1.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
coverage==7.4.4
crispy-bootstrap4==2024.1
cycler==0.12.1
dill==0.3.8
Django==4.2.10
django-bootstrap-v5==1.0.11
django-colorfield==0.11.0
django-crispy-forms==2.1
django-guardian==2.4.0
django-matplotlib==0.1
fonttools==4.51.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
isort==5.13.2
kiwisolver==1.4.5
mando==0.7.1
matplotlib==3.8.4
mccabe==0.7.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
platformdirs==4.2.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pylint==3.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
radon==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tomlkit==0.12.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
